requirements.txt
python-telegram-bot==20.7
feedparser==6.0.11
apscheduler==3.10.4
pytz==2023.3
requests==2.31.0